+++ R/S3-methods.R
@@ -1606,3 +1606,65 @@ print.sj_pval <- function(x, digits = 3, summary = FALSE, ...) {
   x <- purrr::map_if(x, is.numeric, round, digits = digits)
   print.data.frame(as.data.frame(x), ..., row.names = TRUE)
 }
+
+
+#' @export
+print.sj_error_rate <- function(x, digits = 2, ...) {
+  cat(crayon::blue("# Error Rate of Logistic Regression Model\n\n"))
+  cat(sprintf("  Full model: %.2f%%\n", 100 * x$error.model))
+  cat(sprintf("  Null model: %.2f%%\n", 100 * x$error.null))
+}
+
+
+#' @importFrom rlang .data
+#' @export
+print.sj_binres <- function(x, ...) {
+  if (!requireNamespace("ggplot2", quietly = TRUE)) {
+    stop("Package `ggplot2` required.", call. = F)
+  }
+
+  if (!requireNamespace("scales", quietly = TRUE)) {
+    stop("Package `scales` required.", call. = F)
+  }
+
+  x$se.lo <- -x$se
+  if (length(unique(x$group)) > 1)
+    ltitle <- "Within error bounds"
+  else
+    ltitle <- NULL
+
+  term <- attr(x, "term", exact = TRUE)
+  term.label <- attr(x, "term.label", exact = TRUE)
+
+  if (is.null(term))
+    xtitle <- sprintf("Estimated Probability of %s", attr(x, "resp_var", exact = TRUE))
+  else
+    xtitle = term.label
+
+  p <- ggplot2::ggplot(data = x, ggplot2::aes_string(x = "xbar")) +
+    ggplot2::geom_ribbon(ggplot2::aes_string(ymin = -Inf, ymax = "se.lo"), alpha = .1 , fill = "grey70") +
+    ggplot2::geom_ribbon(ggplot2::aes_string(ymin = "se", ymax = Inf), alpha = .1 , fill = "grey70") +
+    ggplot2::geom_line(ggplot2::aes_string(y = "se"), colour = "grey70") +
+    ggplot2::geom_line(ggplot2::aes_string(y = "se.lo"), colour = "grey70") +
+    ggplot2::geom_abline(slope = 0, intercept = 0, colour = "grey80") +
+    ggplot2::theme_bw() +
+    ggplot2::scale_color_manual(values = c("#d11141", "#00aedb")) +
+    ggplot2::labs(
+      y = "Average residual",
+      x = xtitle,
+      colour = ltitle
+    ) +
+    ggplot2::coord_flip()
+
+  if (is.null(term)) {
+    p <- p + ggplot2::scale_x_continuous(labels = scales::percent)
+  }
+
+  if (is.null(ltitle)) {
+    p <- p + ggplot2::geom_point(ggplot2::aes_string(y = "ybar"))
+  } else {
+    p <- p + ggplot2::geom_point(ggplot2::aes_string(y = "ybar", colour = "group"))
+  }
+
+  graphics::plot(p)
+}