+++ R/merMod_p.R
@@ -3,12 +3,11 @@
 #'
 #' @description This function returns the p-values for fitted model objects.
 #'
-#' @param fit A fitted model object of class \code{lm}, \code{glm}, \code{merMod},
-#'        \code{merModLmerTest}, \code{pggls} or \code{gls}. Other classes may
-#'        work as well.
+#' @param fit A model object.
 #' @param p.kr Logical, if \code{TRUE}, the computation of p-values is based on
 #'         conditional F-tests with Kenward-Roger approximation for the df (see
 #'         'Details').
+#' @param ... Currently not used.
 #'
 #' @return A \code{data.frame} with the model coefficients' names (\code{term}),
 #'         p-values (\code{p.value}) and standard errors (\code{std.error}).
@@ -55,84 +54,76 @@
 #' @importFrom stats coef pnorm
 #' @importFrom broom tidy
 #' @importFrom dplyr select
+#' @importFrom stats coef pt pnorm
 #' @export
-p_value <- function(fit, p.kr = FALSE) {
-  # retrieve sigificance level of independent variables (p-values)
-  if (inherits(fit, "pggls")) {
-    p <- summary(fit)$CoefTable[, 4]
-    se <- summary(fit)$CoefTable[, 2]
-  } else if (inherits(fit, "gls")) {
-    p <- summary(fit)$tTable[, 4]
-    se <- summary(fit)$tTable[, 2]
-  } else if (inherits(fit, "Zelig-relogit")) {
+p_value <- function(fit, ...) {
+  UseMethod("p_value")
+}
+
+#' @export
+p_value.multinom <- function(fit, ...) {
+  fit %>%
+    broom::tidy() %>%
+    dplyr::select(.data$term, .data$p.value, .data$std.error)
+}
+
+#' @export
+p_value.polr <- function(fit, ...) {
+  smry <- suppressMessages(as.data.frame(stats::coef(summary(fit))))
+  tstat <- smry[[3]]
+  se <- smry[[2]]
+  p <- 2 * stats::pnorm(abs(tstat), lower.tail = FALSE)
+  names(p) <- rownames(smry)
+  make_it_so(p, se)
+}
+
+#' @export
+p_value.gam <- function(fit, ...) {
+  sm <- summary(fit)
+  lc <- length(sm$p.coeff)
+  p <- sm$p.pv[1:lc]
+  se <- sm$se[1:lc]
+  make_it_so(p, se)
+}
+
+#' @export
+p_value.maxLik <- function(fit, ...) {
+  p <- summary(fit)$estimate[, 4]
+  se <- summary(fit)$estimate[, 2]
+  make_it_so(p, se)
+}
+
+#' @export
+p_value.pglm <- function(fit, ...) {
+  p <- summary(fit)$estimate[, 4]
+  se <- summary(fit)$estimate[, 2]
+  make_it_so(p, se)
+}
+
+#' @export
+p_value.default <- function(fit, ...) {
+  if (grepl("^Zelig-", class(fit))) {
     if (!requireNamespace("Zelig", quietly = T))
       stop("Package `Zelig` required. Please install", call. = F)
     p <- unlist(Zelig::get_pvalue(fit))
     se <- unlist(Zelig::get_se(fit))
-  } else if (is_merMod(fit)) {
-    p <- merMod_p(fit, p.kr)
-    sekr <- attr(p, "se.kr", exact = TRUE)
-    if (!is.null(sekr))
-      se <- sekr
-    else
-      se <- stats::coef(summary(fit))[, 2]
-  } else if (inherits(fit, c("pglm", "maxLik"))) {
-    p <- summary(fit)$estimate[, 4]
-    se <- summary(fit)$estimate[, 2]
-  } else if (inherits(fit, "gam")) {
-    sm <- summary(fit)
-    lc <- length(sm$p.coeff)
-    p <- sm$p.pv[1:lc]
-    se <- sm$se[1:lc]
-  } else if (inherits(fit, "polr")) {
-    smry <- suppressMessages(as.data.frame(stats::coef(summary(fit))))
-    tstat <- smry[[3]]
-    se <- smry[[2]]
-    p <- 2 * stats::pnorm(abs(tstat), lower.tail = FALSE)
-    names(p) <- rownames(smry)
-  } else if (inherits(fit, "multinom")) {
-    return(fit %>%
-      broom::tidy() %>%
-      dplyr::select(.data$term, .data$p.value, .data$std.error))
   } else {
     p <- stats::coef(summary(fit))[, 4]
     se <- stats::coef(summary(fit))[, 2]
   }
-
-  res <- data.frame(
-    term = names(p),
-    p.value = as.vector(p),
-    std.error = as.vector(se)
-  )
-
-  class(res) <- c("sj_pval", class(res))
-
-  attr(res, "df.kr") <- attr(p, "df.kr", exact = TRUE)
-  attr(res, "se.kr") <- attr(p, "se.kr", exact = TRUE)
-  attr(res, "t.kr") <- attr(p, "t.kr", exact = TRUE)
-
-  res
+  make_it_so(p, se)
 }
 
-
-
-
-#' @importFrom stats coef pt pnorm
-merMod_p <- function(fit, p.kr) {
+#' @rdname p_value
+#' @export
+p_value.lmerMod <- function(fit, p.kr = FALSE, ...) {
   # retrieve sigificance level of independent variables (p-values)
   cs <- stats::coef(summary(fit))
 
   # remeber coef-names
   coef_names <- rownames(cs)
 
-  # check if we have p-values in summary
-  if (ncol(cs) >= 4) {
-    # do we have a p-value column?
-    pvcn <- which(colnames(cs) == "Pr(>|t|)")
-    # if not, default to 4
-    if (length(pvcn) == 0) pvcn <- 4
-    pv <- cs[, pvcn]
-  } else if (inherits(fit, "lmerMod") && requireNamespace("pbkrtest", quietly = TRUE) && p.kr) {
+  if (isTRUE(p.kr) && requireNamespace("pbkrtest", quietly = TRUE)) {
     # compute Kenward-Roger-DF for p-statistic. Code snippet adapted from
     # http://mindingthebrain.blogspot.de/2014/02/three-ways-to-get-parameter-specific-p.html
     message("Computing p-values via Kenward-Roger approximation. Use `p.kr = FALSE` if computation takes too long.")
@@ -143,23 +134,152 @@ merMod_p <- function(fit, p.kr) {
     se.kr <- get_kr_se(fit)
     t.kr <- lme4::fixef(fit) / se.kr
     # compute p-values, assuming an approximate t-dist
-    pv <- 2 * stats::pt(abs(t.kr), df = df.kr, lower.tail = FALSE)
+    p <- 2 * stats::pt(abs(t.kr), df = df.kr, lower.tail = FALSE)
     # name vector
-    names(pv) <- coef_names
-    attr(pv, "df.kr") <- df.kr
-    attr(pv, "se.kr") <- se.kr
-    attr(pv, "t.kr") <- t.kr
+    names(p) <- coef_names
+    attr(p, "df.kr") <- df.kr
+    attr(p, "se.kr") <- se.kr
+    attr(p, "t.kr") <- t.kr
   } else {
     message("Computing p-values via Wald-statistics approximation (treating t as Wald z).")
-    pv <- 2 * stats::pnorm(abs(cs[, 3]), lower.tail = FALSE)
+    p <- 2 * stats::pnorm(abs(cs[, 3]), lower.tail = FALSE)
   }
 
-
-  if (is.null(names(pv))) {
-    if (length(coef_names) == length(pv)) names(pv) <- coef_names
+  if (is.null(names(p))) {
+    if (length(coef_names) == length(p)) names(p) <- coef_names
   }
 
-  pv
+  sekr <- attr(p, "se.kr", exact = TRUE)
+  if (!is.null(sekr))
+    se <- sekr
+  else
+    se <- cs[, 2]
+
+  make_it_so(p, se)
+}
+
+#' @export
+p_value.glmerMod <- function(fit, ...) {
+  cs <- stats::coef(summary(fit))
+  x <- check_pval_from_summary(fit, cs)
+  if (is.null(x[[1]])) x <- wald_pval(fit, cs)
+
+  make_it_so(x$p, x$se)
+}
+
+#' @importFrom purrr compact map flatten_chr map_df
+#' @importFrom dplyr bind_cols
+#' @export
+p_value.glmmTMB <- function(fit, ...) {
+  cs <- purrr::compact(stats::coef(summary(fit)))
+  x <- purrr::map(cs, ~ check_pval_from_summary(fit, .x))
+  coef_names <- purrr::map(x, ~ names(.x[[1]])) %>% purrr::flatten_chr()
+  model_names <- names(x)
+  model_lengths <- length(x)
+  x <- purrr::map_df(x, ~ dplyr::bind_cols(.x))
+  p <- x[[1]]
+  se <- x[[2]]
+  names(p) <- coef_names
+
+  make_it_so(p, se, rep(model_names, each = nrow(x) / model_lengths))
+}
+
+#' @export
+p_value.nlmerMod <- function(fit, ...) {
+  cs <- stats::coef(summary(fit))
+  x <- check_pval_from_summary(fit, cs)
+  if (is.null(x[[1]])) x <- wald_pval(fit, cs)
+
+  make_it_so(x$p, x$se)
+}
+
+#' @export
+p_value.lmerModLmerTest <- function(fit, ...) {
+  cs <- stats::coef(summary(fit))
+  x <- check_pval_from_summary(fit, cs)
+  if (is.null(x[[1]])) x <- wald_pval(fit, cs)
+
+  make_it_so(x$p, x$se)
+}
+
+#' @export
+p_value.gls <- function(fit, ...) {
+  p <- summary(fit)$tTable[, 4]
+  se <- summary(fit)$tTable[, 2]
+  make_it_so(p, se)
+}
+
+#' @export
+p_value.pggls <- function(fit, ...) {
+  p <- summary(fit)$CoefTable[, 4]
+  se <- summary(fit)$CoefTable[, 2]
+  make_it_so(p, se)
+}
+
+
+#' @importFrom stats coef
+check_pval_from_summary <- function(fit, cs) {
+  p <- NULL
+  se <- NULL
+
+  if (ncol(cs) >= 4) {
+
+    # do we have a p-value column based on t?
+    pvcn <- which(colnames(cs) == "Pr(>|t|)")
+
+    # if not, do we have a p-value column based on z?
+    if (length(pvcn) == 0)
+      pvcn <- which(colnames(cs) == "Pr(>|z|)")
+
+    # if not, default to 4
+    if (length(pvcn) == 0) pvcn <- 4
+
+    p <- cs[, pvcn]
+
+    if (is.null(names(p))) {
+      coef_names <- rownames(cs)
+      if (length(coef_names) == length(p)) names(p) <- coef_names
+    }
+
+    se <- cs[, 2]
+  }
+
+  list(p, se)
+}
+
+
+wald_pval <- function(fit, cs) {
+  message("Computing p-values via Wald-statistics approximation (treating t as Wald z).")
+
+  p <- 2 * stats::pnorm(abs(cs[, 3]), lower.tail = FALSE)
+  se <- cs[, 2]
+
+  coef_names <- rownames(cs)
+
+  if (is.null(names(p))) {
+    if (length(coef_names) == length(p)) names(p) <- coef_names
+  }
+
+  list(p, se)
+}
+
+
+make_it_so <- function(p, se, model = NULL) {
+  res <- data.frame(
+    term = names(p),
+    p.value = as.vector(p),
+    std.error = as.vector(se)
+  )
+
+  if (!is.null(model)) res$model <- model
+
+  class(res) <- c("sj_pval", class(res))
+
+  attr(res, "df.kr") <- attr(p, "df.kr", exact = TRUE)
+  attr(res, "se.kr") <- attr(p, "se.kr", exact = TRUE)
+  attr(res, "t.kr") <- attr(p, "t.kr", exact = TRUE)
+
+  res
 }